Moved views to their own class and file
