+++ olGO/Views/Auth/LogoutView.swift
@@ -1,0 +1,55 @@
+//
+//  LogoutView.swift
+//  olGO
+//
+//  Created by Zach Eriksen on 11/29/19.
+//  Copyright © 2019 oneleif. All rights reserved.
+//
+
+import UIKit
+import SwiftUIKit
+import Combine
+
+class LogoutView: UIView {
+    private var bag: [AnyCancellable] = []
+    
+    private var isRequesting: Bool = false
+    
+    init() {
+        super.init(frame: .zero)
+        backgroundColor = .white
+        embed {
+            Button("Logout", titleColor: .blue) {
+                self.logout()
+            }
+            .padding()
+        }
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    private func logout() {
+        guard !self.isRequesting else {
+            return
+        }
+        self.isRequesting = true
+        self.bag.append(API.instance.logout()
+            .sink(receiveCompletion: { (result) in
+                
+                self.isRequesting = false
+                if case .failure(let error) = result {
+                    print(error.localizedDescription)
+                }
+                
+            }) { (data, response) in
+                
+                guard let response = response as? HTTPURLResponse else {
+                    return
+                }
+                
+                print("Logout Response Status Code: \(response.statusCode)")
+        })
+    }
+}
