+++ olGO/Views/Auth/RegisterView.swift
@@ -14,7 +14,7 @@ class RegisterView: UIView {
     private var username: String = ""
     private var password: String = ""
     private var bag: [AnyCancellable] = []
-
+    
     private var isRequesting: Bool = false
     
     init() {
@@ -42,27 +42,7 @@ class RegisterView: UIView {
                         ]
                     },
                     Button("Register", titleColor: .blue) {
-                        guard !self.isRequesting else {
-                            return
-                        }
-                        self.isRequesting = true
-                        self.bag.append(API.instance.register(user: User(username: self.username,
-                                                                         password: self.password))
-                            .sink(receiveCompletion: { (result) in
-                                self.isRequesting = false
-                                switch result {
-                                case .failure(let error):
-                                    print(error.localizedDescription)
-                                case .finished:
-                                    DispatchQueue.main.async {
-                                        Navigate.shared.go(UIViewController {
-                                                             LoginView()
-                                        }, style: .push)
-                                    }
-                                }
-                            }) { (data, response) in
-                                print("Did Register!")
-                        })
+                        self.register()
                     }
                     .frame(height: 60),
                     Spacer()
@@ -74,6 +54,36 @@ class RegisterView: UIView {
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
+    }
+    
+    private func register() {
+        guard !self.isRequesting else {
+            return
+        }
+        self.isRequesting = true
+        self.bag.append(API.instance.register(user: User(username: self.username,
+                                                         password: self.password))
+            .sink(receiveCompletion: { (result) in
+                self.isRequesting = false
+                if case .failure(let error) = result {
+                    print(error.localizedDescription)
+                }
+            }) { (data, response) in
+                guard let response = response as? HTTPURLResponse else {
+                    return
+                }
+                
+                print("Register Response Status Code: \(response.statusCode)")
+                
+                if 200 ... 300 ~= response.statusCode {
+                    DispatchQueue.main.async {
+                        Navigate.shared.go(UIViewController {
+                            LoginView()
+                        }, style: .push)
+                    }
+                }
+                
+        })
     }
 }
