+++ olGO/Views/Posts/AddPostView.swift
@@ -1,0 +1,9 @@
+//
+//  AddPostView.swift
+//  olGO
+//
+//  Created by Zach Eriksen on 11/30/19.
+//  Copyright © 2019 oneleif. All rights reserved.
+//
+
+import Foundation
