+++ olGO/Views/Social/SocialView.swift
@@ -1,0 +1,9 @@
+//
+//  SocialView.swift
+//  olGO
+//
+//  Created by Zach Eriksen on 11/30/19.
+//  Copyright © 2019 oneleif. All rights reserved.
+//
+
+import Foundation
